==> combine_effort/1012. 21-march-2021 topredict 28-march-2021/4. predict_nextweek.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import RandomForestRegressor
from sklearn.neighbors import KNeighborsClassifier
from sklearn.neural_network import MLPClassifier
#import matplotlib.pyplot as plt
from sklearn import metrics
import numpy as np
import ast
import time
from datetime import datetime

#currency_pairs = ['USDCHF10080',
                 #'GBPUSD10080', 'EURUSD10080', 'USDJPY10080', 'USDCAD10080', 'AUDUSD10080', 'NZDUSD10080',
                 #'GBPCHF10080', 'EURCHF10080', 'CHFJPY10080', 'CADCHF10080', 'AUDCHF10080', 'NZDCHF10080', 'EURGBP10080',
                 #'GBPJPY10080', 'GBPCAD10080', 'GBPAUD10080', 'EURJPY10080', 'EURCAD10080', 'EURAUD10080', 'EURNZD10080',
                #'CADJPY10080', 'AUDJPY10080', 'NZDJPY10080', 'AUDCAD10080', 'NZDCAD10080', 'AUDNZD10080'
                #]

headers = ['date', 'ignore', 'open', 'high', 'low', 'close', 'volume'  ]


q = 0
last_trial_time = datetime.now()
while q < 1:
    
    print(last_trial_time)
    control_dataframe = pd.read_csv('control.csv' ) 
    
    if control_dataframe['feature'].iloc[:1].values == 1:     
        
        q = q+1
        
        
        
    else:  
        last_trial_time = datetime.now()
        time.sleep(60)
        

print('---------------- Script Has Started Execution-----------------------')

next_week_prediction_score = []
next_week_prediction =[]

next_week_trades = pd.read_csv('files/next_week_trades.csv'  ) 




for index, row in next_week_trades.iterrows():
    
   
    
    current_currency = row.names
    raw_data = pd.read_csv('files/' + current_currency + '.csv'  ) 
    weeks = row.week
    wanted_features = ast.literal_eval(row.needed_features)
    n_estimators = 20000
    min_samples_leaf = 1
    max_features = "auto"
    max_samples = None




    raw_data = raw_data[:weeks]

    X_train = pd.DataFrame()

    for q in wanted_features:

        X_train[q] = raw_data[q]


    #print(X_train)  

    y_train = raw_data['nextweek_'+ current_currency]

    #print(y_train)



    prediction_model=RandomForestRegressor(
                                                n_estimators=n_estimators,
                                               n_jobs =-1,
                                               min_samples_leaf=min_samples_leaf,
                                                max_features=max_features,
                                               max_samples = max_samples
                                          )



    prediction_model.fit(X_train,y_train)


    current_week_values = pd.read_csv('current_week_values/current_week_values.csv'  ) 

    X_pred = pd.DataFrame()

    for h in wanted_features:

        X_pred[h] = current_week_values[h]

    #print(X_pred)

    y_pred = prediction_model.predict(X_pred)

    
    if y_pred[0] > 0:
        
        value = 'B'
        
    else:
        value = 'S'
        
    next_week_prediction_score.append(y_pred[0])
    next_week_prediction.append(value)
    
    print(row.names, row.week, y_pred)



print(next_week_prediction)
    

next_week_trades['next_week_prediction_score'] = next_week_prediction_score
next_week_trades['next_week_prediction'] = next_week_prediction

next_week_trades.to_csv("files/next_week_trades.csv")

control_dataframe = pd.read_csv('control.csv' ) 
control_dataframe['predict'].iloc[:1] = 1


print(control_dataframe)
control_dataframe.to_csv('control.csv', index=False)
